% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gpt_helpers.R
\name{add_message}
\alias{add_message}
\title{Message settings}
\usage{
add_message(config, thread_id, message, files = NULL)
}
\arguments{
\item{config}{The configuration object}

\item{thread_id}{The thread ID}

\item{message}{The message to add}

\item{files}{Optional list of files to attach}
}
\value{
Message content
}
\description{
Message settings
}
