% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gpt_helpers.R
\name{create_config}
\alias{create_config}
\title{Create configuration for API calls}
\usage{
create_config(api_key, assistant_id)
}
\arguments{
\item{api_key}{The OpenAI API key}

\item{assistant_id}{The assistant ID}
}
\value{
A list containing configuration parameters
}
\description{
Create configuration for API calls
}
