% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gpt_helpers.R
\name{create_thread}
\alias{create_thread}
\title{Create a new thread}
\usage{
create_thread(config)
}
\arguments{
\item{config}{The configuration object}
}
\value{
Thread ID
}
\description{
Create a new thread
}
