% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers_server_to_template.R
\name{extract_and_generate_code}
\alias{extract_and_generate_code}
\title{Extract all settings and generate complete implementation code}
\usage{
extract_and_generate_code(input)
}
\arguments{
\item{input}{Shiny input object}
}
\value{
List of code components
}
\description{
Extract all settings and generate complete implementation code
}
