% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_anotheRgenai_api.R
\name{fetch_assistants}
\alias{fetch_assistants}
\title{List available OpenAI assistants}
\usage{
fetch_assistants(config)
}
\arguments{
\item{config}{Fetch assistants protocole with GET}
}
\value{
Data frame of assistants
}
\description{
List available OpenAI assistants
}
