% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_anotheRgenai_api.R
\name{fetch_response}
\alias{fetch_response}
\title{Retrieve assistant response}
\usage{
fetch_response(config, thread_id)
}
\arguments{
\item{config}{The enhanced configuration object}

\item{thread_id}{The thread ID}
}
\value{
The message text from the assistant's response
}
\description{
Retrieve assistant response
}
