% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dynamic_gpt_agent_server.R
\name{generate_chat_customization_ui}
\alias{generate_chat_customization_ui}
\title{Generate chat interface customization UI}
\usage{
generate_chat_customization_ui(ns)
}
\arguments{
\item{ns}{Namespace function for the Shiny module}
}
\value{
Shiny UI elements for chat customization
}
\description{
Generate chat interface customization UI
}
