% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers_server_to_template.R
\name{generate_css_styles}
\alias{generate_css_styles}
\title{Generate CSS styles from settings}
\usage{
generate_css_styles(settings)
}
\arguments{
\item{settings}{Extracted UI settings}
}
\value{
String containing CSS styles
}
\description{
Generate CSS styles from settings
}
