% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_templates.R
\name{generate_implementation_code}
\alias{generate_implementation_code}
\title{Template-based implementation code generator}
\usage{
generate_implementation_code(settings, input)
}
\arguments{
\item{settings}{List of settings for chat interface}

\item{input}{Shiny input object for UI customizations}
}
\value{
List of code sections
}
\description{
Template-based implementation code generator
}
