% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dynamic_gpt_agent_server.R
\name{generate_implementation_tab}
\alias{generate_implementation_tab}
\title{Add a new tab to the UI for implementation}
\usage{
generate_implementation_tab(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
tabPanel ready to go
}
\description{
Add a new tab to the UI for implementation
}
