% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers_server_to_template.R
\name{generate_message_renderer_code}
\alias{generate_message_renderer_code}
\title{Generate message rendering code}
\usage{
generate_message_renderer_code(settings)
}
\arguments{
\item{settings}{Extracted UI settings}
}
\value{
String containing message rendering code
}
\description{
Generate message rendering code
}
