% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers_server_to_template.R
\name{generate_ui_component_code}
\alias{generate_ui_component_code}
\title{Generate UI component code}
\usage{
generate_ui_component_code(settings)
}
\arguments{
\item{settings}{Extracted UI settings}
}
\value{
String containing UI component code
}
\description{
Generate UI component code
}
