% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dynamic_gpt_agent_server.R
\name{get_color_scheme}
\alias{get_color_scheme}
\title{Get color scheme for chat bubbles}
\usage{
get_color_scheme(theme_name)
}
\arguments{
\item{theme_name}{Name of the color theme}
}
\value{
List containing user and assistant colors
}
\description{
Get color scheme for chat bubbles
}
