% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gpt_helpers.R
\name{run_assistant}
\alias{run_assistant}
\title{Add message to thread}
\usage{
run_assistant(config, thread_id)
}
\arguments{
\item{config}{The configuration object}

\item{thread_id}{The thread ID}
}
\value{
Message content
}
\description{
Add message to thread
}
